+++ tests/provider/dwd/observation/test_api_metadata.py
@@ -14,26 +14,22 @@ from wetterdienst.provider.dwd.observation import (
 def test_dwd_observation_metadata_discover_parameters():
     parameters = DwdObservationRequest.discover(filter_="minute_1", flatten=True)
 
-    assert (
-        json.dumps(
-            {
-                "minute_1": {
-                    "precipitation_height": {"origin": "mm", "si": "kg / m ** 2"},
-                    "precipitation_height_droplet": {
-                        "origin": "mm",
-                        "si": "kg / m ** 2",
-                    },
-                    "precipitation_height_rocker": {
-                        "origin": "mm",
-                        "si": "kg / m ** 2",
-                    },
-                    "precipitation_form": {"origin": "-", "si": "-"},
-                }
+    expected = {
+        "minute_1": {
+            "precipitation_height": {"origin": "mm", "si": "kg / m ** 2"},
+            "precipitation_height_droplet": {
+                "origin": "mm",
+                "si": "kg / m ** 2",
             },
-            indent=4,
-        )
-        in parameters
-    )
+            "precipitation_height_rocker": {
+                "origin": "mm",
+                "si": "kg / m ** 2",
+            },
+            "precipitation_form": {"origin": "-", "si": "-"},
+        }
+    }
+
+    assert json.dumps(expected) in json.dumps(parameters)
 
 
 def test_dwd_observation_metadata_describe_fields_kl_daily_english():